requirements/requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==21.11.1
Automat==20.2.0
bcrypt==3.2.0
black==21.12b0
boto3==1.20.24
botocore==1.23.27
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.9
click==8.0.3
constantly==15.1.0
cryptography==36.0.0
daphne==3.0.2
Django==3.2.8
django-cors-headers==3.10.0
django-crontab==0.7.1
django-extensions==3.1.5
django-filter==21.1
django-storages==1.12.3
djangorestframework==3.13.1
drf-nested-routers==0.93.4
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.10.0
incremental==21.3.0
jmespath==0.10.0
Markdown==3.3.6
msgpack==1.0.3
mypy-extensions==0.4.3
mysqlclient==2.0.3
pathspec==0.9.0
platformdirs==2.4.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
pyOpenSSL==21.0.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
s3transfer==0.5.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tomli==1.2.3
Twisted==21.7.0
txaio==21.2.1
typing_extensions==4.0.1
urllib3==1.26.7
zipp==3.7.0
zope.interface==5.4.0
